refactor : 댓글 기능 수정

+++ post-service/src/main/java/com/todaytrend/postservice/post/repository/PostLikeRepository.java
@@ -16,5 +16,5 @@ public interface PostLikeRepository extends JpaRepository<PostLike,Long>{
 
     void deleteByUserUuidAndPostId(String userUuid, Long postId);
 
-    int countByPostId(Long postId);
+    Long countByPostId(Long postId);
 }
